feat: fix register order, register withdrawal

+++ internal/migrations/20240915073004_new_users_table.sql
@@ -5,7 +5,7 @@ CREATE TABLE IF NOT EXISTS users
     id serial PRIMARY KEY,
     login VARCHAR (100) NOT NULL unique,
     hash VARCHAR (100) NOT NULL,
-    balance BIGINT NOT NULL DEFAULT 0,
+    balance DOUBLE PRECISION NOT NULL DEFAULT 0,
     created_at timestamp without time zone NOT NULL DEFAULT (current_timestamp AT TIME ZONE 'UTC'),
     updated_at timestamp without time zone NOT NULL DEFAULT (current_timestamp AT TIME ZONE 'UTC')
     );
